remove timeout parameter from open_serial
The method open_serial could be called several times in the lifecycle of
the class but we might not want to change the recieve timeout, nor reset
it to the default value. Therefore, it is initialised by the constructor
(isn't it cleaner anyway ?) and still accessible with a getter and a
setter.