Allow to change timeout for scanning of servos.
In the command-line utility, most commands require that first the
available servos are searched on the bus. For faster search, a fixed,
small, timeout was used.

We observed that this value could bring problems for some serial
interfaces. Hence, it is now possible to manually override this timeout,
through the new option "-scan-timeout".